requirements.txt
# Speech Analyzer - Python Dependencies
# AI-Powered Voice Assessment Tool

# Core Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# Speech-to-Text
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Natural Language Processing
transformers>=4.30.0
tokenizers>=0.13.0
huggingface-hub>=0.16.0
nltk>=3.8
textstat>=0.7.3

# Data Science and Analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
datasets>=2.14.0

# Utilities
click>=8.0.0
colorama>=0.4.6
tqdm>=4.65.0
requests>=2.31.0

# Model Training (optional - for advanced training)
accelerate>=0.21.0
wandb>=0.15.0
tensorboard>=2.13.0

# Optional: GPU acceleration (uncomment if you have CUDA-compatible GPU)
# torch>=2.0.0+cu118
# torchaudio>=2.0.0+cu118

# Development and Testing (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
joblib>=1.3.0

# Documentation (optional)
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
